% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_points.R
\name{color_points}
\alias{color_points}
\title{Define colors based on coordinates}
\usage{
color_points(
  x,
  y,
  cx = 1,
  ex = 1,
  cy = 0,
  ey = 1,
  ntl = 300,
  plt = c("#FFFAAA", "#EECC00", "#DD5500", "#993300", "#550000")
)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric}

\item{cx}{Numeric}

\item{ex}{Numeric}

\item{cy}{Numeric}

\item{ey}{range Numeric}

\item{ntl}{Numeric}

\item{plt}{Character vector}
}
\value{
ggplot color
}
\description{
Function
}
\author{
Nicolas Mangin
}
