% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_polar_parameters.R
\name{gather_polar_parameters}
\alias{gather_polar_parameters}
\title{Collect and process polar parameters}
\usage{
gather_polar_parameters(path = "C:/Users/nicol/Pictures/Art/Polar")
}
\arguments{
\item{path}{Character. Path to the folder containing parameter files.}
}
\description{
Module collecting the paraleters used to produce a polar distorsion.
}
\author{
Nicolas Mangin
}
