% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_parameters.R
\docType{data}
\name{polar_parameters}
\alias{polar_parameters}
\title{polar_parameters}
\format{
A data frame with 39 rows and 24 variables
\itemize{
\item  name   character
\item  file   character
\item  a   character
\item  b   character
\item  c   character
\item  d   character
\item  gamma   character
\item  range   character
\item  density   character
\item  x   character
\item  y   character
\item  seed   character
\item  alp   character
\item  sz   character
\item  shp   character
\item  plt   character
\item  bkgcol   character
\item  coord   character
\item  cx   character
\item  ex   character
\item  cy   character
\item  ey   character
\item  ntl   character
\item  rotation   character
}
}
\usage{
data(polar_parameters)
}
\description{
polar_parameters
}
\keyword{datasets}
