% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_server.R
\name{polar_server}
\alias{polar_server}
\title{Collect and process polar parameters}
\usage{
polar_server(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\description{
Module collecting the paraleters used to produce a polar distorsion.
}
\author{
Nicolas Mangin
}
