% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_ui.R, R/rosas_ui.R
\name{polar_ui}
\alias{polar_ui}
\title{Collect and process polar parameters}
\usage{
polar_ui(id)

polar_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\description{
Module collecting the paraleters used to produce a polar distorsion.

Module collecting the paraleters used to produce a polar distorsion.
}
\author{
Nicolas Mangin
}
