% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flametree_server.R, R/rosas_server.R
\name{rosas_server}
\alias{rosas_server}
\title{Collect and process polar parameters}
\usage{
rosas_server(id)

rosas_server(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\description{
Module collecting the paraleters used to produce a polar distorsion.

Module collecting the paraleters used to produce a polar distorsion.
}
\author{
Nicolas Mangin
}
