% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtist_help.R
\name{rtist_help}
\alias{rtist_help}
\title{Name and painting for all rtist palettes}
\usage{
rtist_help()
}
\value{
A printed list of rtist palettes with name of their associated painting and artist.
}
\description{
Use function to get a print of all palettes with painting and artist name.
}
\examples{
rtist_help()
}
