% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rtist_palette}
\alias{rtist_palette}
\title{A generator for color palettes from famous artists}
\usage{
rtist_palette(artist, n, type = c("discrete", "continuous"))
}
\arguments{
\item{artist}{Name of desired palette.}

\item{n}{Number of colors desired. All palettes now only have five colors.
All color schemes are taken from this blog article:
\href{http://www.skellermeyerdesigns.com/blog/10-color-palettes-based-on-famous-paintings}{10 Color Palettes Based on Famous Paintings}.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are some color palettes from famous works of art by well known artist.
}
\examples{
rtist_palette("vangogh")
rtist_palette("hokusai", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- rtist_palette(21, artist = "raphael", type = "continuous")
}
\keyword{colors}
