% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_art.R
\name{save_art}
\alias{save_art}
\title{Select and load a intake}
\usage{
save_art(
  art,
  name = "plot",
  format = ".png",
  folder = base::getwd(),
  wdth = 95,
  hght = 95
)
}
\arguments{
\item{art}{ggplot graph.}

\item{name}{Character.}

\item{format}{Character}

\item{folder}{Character}

\item{wdth}{Numeric}

\item{hght}{Numeric}
}
\value{
A list containing the intake as a table and a json object
}
\description{
Module facilitating the loading of a classification intake.
}
\author{
Nicolas Mangin
}
