% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_distorsion.R
\name{polar_distorsion}
\alias{polar_distorsion}
\title{Draw a polar distorsion}
\usage{
polar_distorsion(
  a = 2,
  b = 2,
  c = 3,
  d = 2,
  gamma = 2,
  range = 5,
  density = 50,
  x = 1,
  y = -1,
  seed = 20240713,
  alp = 0.05,
  sz = 0.75,
  shp = 16,
  drwcol = grDevices::hsv(0.4, 0.4, 1),
  bkgcol = grDevices::hsv(0.6, 0.5, 0.2, 1),
  rotation = 0,
  coord = "polar"
)
}
\arguments{
\item{a}{Numeric}

\item{b}{Numeric}

\item{c}{Numeric}

\item{d}{Numeric}

\item{gamma}{Numeric}

\item{range}{Numeric}

\item{density}{Numeric}

\item{x}{Numeric}

\item{y}{Numeric}

\item{seed}{Numeric}

\item{alp}{Numeric}

\item{sz}{Numeric}

\item{shp}{Numeric}

\item{drwcol}{Character}

\item{bkgcol}{Character}

\item{rotation}{Numeric. angle of rotation}

\item{coord}{Character. nothing, "polar" or "radial"}
}
\value{
ggplot graph.
}
\description{
Module
}
\author{
Nicolas Mangin
}
