% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flametree.R, R/rosas.R
\name{rosas}
\alias{rosas}
\alias{flame_trees}
\title{Draw a rosas}
\usage{
flame_trees()

rosas(
  range = 0.5,
  density = 5000,
  colors = base::rep(c("yellow", "orange", "red", "firebrick"), 2),
  curv = 0.5,
  size = 1,
  alpha = 0.01,
  background = "#330000",
  limx = 3,
  limy = 3,
  coord = "equal"
)
}
\arguments{
\item{range}{Numeric}

\item{density}{Numeric}

\item{colors}{Character vector}

\item{curv}{Numeric}

\item{size}{Numeric}

\item{alpha}{Numeric}

\item{background}{Character}

\item{limx}{Numeric}

\item{limy}{Numeric}

\item{coord}{Character. "equal", "polar" or "radial"}
}
\value{
ggplot graph.

ggplot graph.
}
\description{
Module

Module
}
\author{
Nicolas Mangin
}
