% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{rtist_palettes}
\alias{rtist_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 10.}
\usage{
rtist_palettes
}
\description{
Use \code{\link{rtist_palette}} to construct palettes of desired length.
}
\keyword{datasets}
